add profile_referee_delete: It works!!!

+++ accounts/views.py
@@ -1,6 +1,6 @@
 from django.db.transaction import atomic
 from django.http import HttpResponseRedirect
-from django.shortcuts import get_object_or_404
+from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse, reverse_lazy
 from django.views.generic import CreateView, UpdateView, DeleteView
 
@@ -11,91 +11,29 @@ from referees.models import Referee, RefereeLicenceType
 
 
 class ProfileRefereeAddView(CreateView):
-    model = Referee
     form_class = AddProfileRefereeForm
-    template_name = "form_add.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['licence_types'] = RefereeLicenceType.objects.all()
-        context['cities'] = City.objects.all()
-        return context
-
-    def form_valid(self, form):
-        referee, user = form.save(commit=True)
-        return HttpResponseRedirect(self.get_success_url())
-
-    def get_success_url(self):
-        return reverse('referees_list')
+    template_name = "form.html"
+    success_url = reverse_lazy('referees_list')
 
 
 class ProfileRefereeEditView(UpdateView):
     model = Referee
     form_class = AddProfileRefereeForm
-    template_name = "form_edit.html"
-
-    def get_object(self, queryset=None):
-        referee_id = self.kwargs.get("pk")
-        return get_object_or_404(Referee, pk=referee_id)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['licence_types'] = RefereeLicenceType.objects.all()
-        context['cities'] = City.objects.all()
-
-        referee = self.get_object()
-        profile = ProfileReferee.objects.get(referee=referee)
-        context['user'] = profile.user
-        return context
-
-    def form_valid(self, form):
-        referee = self.get_object()
-        profile = ProfileReferee.objects.get(referee=referee)
-
-        # Update referee information
-        referee.licence_number = form.cleaned_data['licence_number']
-        referee.licence_type = form.cleaned_data['licence_type']
-        referee.city = form.cleaned_data['city']
-        referee.rating = form.cleaned_data['rating']
-        referee.phone = form.cleaned_data['phone']
-        referee.save()
-
-        # Update user information
-        user = profile.user
-        user.first_name = form.cleaned_data['name']
-        user.last_name = form.cleaned_data['surname']
-        user.email = form.cleaned_data['email']
-        user.save()
-
-        return HttpResponseRedirect(self.get_success_url())
-
-    def get_success_url(self):
-        return reverse('referees_list')
-
-
-class ProfileRefereeDeleteView(DeleteView):
-    model = ProfileReferee
-    template_name = "form_delete.html"
+    template_name = "form.html"
     success_url = reverse_lazy('referees_list')
 
-    @atomic
-    def delete(self, request, *args, **kwargs):
-        profile_referee = self.get_object()
-        referee = profile_referee.referee
-        user = profile_referee.user
 
-        # Deaktivace uživatele
-        user.is_active = False
-        user.save()
-
-        # Smazání rozhodčího, pokud má přiřazený profil
+@atomic
+def profile_referee_delete(request, pk):
+    referee = get_object_or_404(Referee, pk=pk)
+    profile_referee = ProfileReferee.objects.get(referee=referee)
+    user = profile_referee.user
+    if request.method == 'POST':
+        print(profile_referee)
+        referee.profile.delete()
+        print(referee)
         referee.delete()
-
-        # Smazání profilu rozhodčího
-        profile_referee.delete()
-
-        # Kontrola a smazání orphan referee záznamů
-        orphaned_referees = Referee.objects.filter(profile=None)
-        orphaned_referees.delete()
-
-        return HttpResponseRedirect(self.get_success_url())
+        print(user)
+        user.delete()
+        return redirect('referees_list')  # Přejděte na stránku úspěchu
+    return render(request, 'form_delete.html', {'referee': referee})
